// src/enums/tags.ts
//tag of guests 
export enum TagEnglish {
    HIGH_SCHOOL_FRIEND = "High School Friend",
    COLLEGE_FRIEND = "College Friend",
    FRIEND = "Friend",
    TEAMWORK = "Teamwork",
    RELATIVE = "Relative",
    OTHERS = "Others",
}

//tag of guests khmer
export enum TagKhmer {
    HIGH_SCHOOL_FRIEND = "មិត្តភក្តិវិទ្យាល័យ",
    COLLEGE_FRIEND = "មិត្តភក្តិឧត្តមសិក្សា",
    FRIEND = "មិត្តភក្តិ",
    TEAMWORK = "ការងារក្រុម",
    RELATIVE = "សាច់ញាតិ",
    OTHERS = "ផ្សេងៗ",
}